# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Nagios Management Portal API",
  "version": "1.0.8"
 },
 "paths": {
...
  "/api/config/validate": {
   "post": {
    "tags": [
     "configuration"
    ],
    "summary": "Validate",
    "operationId": "validate_api_config_validate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/RunIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/apply": {
   "post": {
    "tags": [
     "configuration"
    ],
    "summary": "Apply",
    "operationId": "apply_api_config_apply_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/RunIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/versions": {
   "get": {
    "tags": [
     "configuration"
    ],
    "summary": "Versions",
    "operationId": "versions_api_config_versions_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 25,
       "title": "Page Size"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 30
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/versions/{vid}": {
   "get": {
    "tags": [
     "configuration"
    ],
    "summary": "Version Detail",
    "operationId": "version_detail_api_config_versions__vid__get",
    "parameters": [
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/versions/{vid}/file": {
   "get": {
    "tags": [
     "configuration"
    ],
    "summary": "Version File",
    "operationId": "version_file_api_config_versions__vid__file_get",
    "parameters": [
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vid"
      }
     },
     {
      "name": "path",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "maxLength": 300,
       "title": "Path"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/versions/{a}/diff/{b}": {
   "get": {
    "tags": [
     "configuration"
    ],
    "summary": "Version Diff",
    "operationId": "version_diff_api_config_versions__a__diff__b__get",
    "parameters": [
     {
      "name": "a",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "A"
      }
     },
     {
      "name": "b",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "B"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/versions/{vid}/download": {
   "get": {
    "tags": [
     "configuration"
    ],
    "summary": "Version Download",
    "operationId": "version_download_api_config_versions__vid__download_get",
    "parameters": [
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/versions/{vid}/apply": {
   "post": {
    "tags": [
     "configuration"
    ],
    "summary": "Apply Existing",
    "operationId": "apply_existing_api_config_versions__vid__apply_post",
    "parameters": [
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/config/versions/{vid}/rollback": {
   "post": {
    "tags": [
     "configuration"
    ],
    "summary": "Rollback",
    "operationId": "rollback_api_config_versions__vid__rollback_post",
    "parameters": [
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Vid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/backups": {
   "get": {
    "tags": [
     "backups"
    ],
    "summary": "List Backups",
    "operationId": "list_backups_api_backups_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "backups"
    ],
    "summary": "Create Backup",
    "operationId": "create_backup_api_backups_post",
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   }
  },
  "/api/backups/{bid}/download": {
   "get": {
    "tags": [
     "backups"
    ],
    "summary": "Download Backup",
    "operationId": "download_backup_api_backups__bid__download_get",
    "parameters": [
     {
      "name": "bid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Bid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/backups/{bid}/restore": {
   "post": {
    "tags": [
     "backups"
    ],
    "summary": "Restore Backup",
    "description": "Emergency file-level restore (does not change the database - portal shows drift afterwards).",
    "operationId": "restore_backup_api_backups__bid__restore_post",
    "parameters": [
     {
      "name": "bid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Bid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/import": {
   "post": {
    "tags": [
     "import"
    ],
    "summary": "Run Import",
    "operationId": "run_import_api_import_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ImportIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "ImportIn": {
    "properties": {
     "mode": {
      "type": "string",
      "enum": [
       "takeover",
       "readonly"
      ],
      "title": "Mode",
      "default": "takeover"
     },
     "hosts": {
      "items": {
       "$ref": "#/components/schemas/ImportSelection"
      },
      "type": "array",
      "maxItems": 500,
      "minItems": 1,
      "title": "Hosts"
     },
     "apply": {
      "type": "boolean",
      "title": "Apply",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "hosts"
    ],
    "title": "ImportIn"
   },
   "ImportSelection": {
    "properties": {
     "hostname": {
      "type": "string",
      "maxLength": 64,
      "title": "Hostname"
     },
     "environment": {
      "type": "string",
      "enum": [
       "production",
       "uat",
       "development",
       "dr",
       "test"
      ],
      "title": "Environment",
      "default": "production"
     },
     "os_type": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "windows",
         "linux",
         "unix",
         "network",
         "other"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Os Type"
     },
     "location": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     }
    },
    "type": "object",
    "required": [
     "hostname"
    ],
    "title": "ImportSelection"
   },
...
   "RunIn": {
    "properties": {
     "summary": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Summary"
     }
    },
    "type": "object",
    "title": "RunIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}